--- src/main/java/com/devkurly/product/pagehandler/PageHandler.java
@@ -1,120 +1,0 @@
-package com.devkurly.product.pagehandler;
-
-public class PageHandler {
-
-    private int totalCnt;
-    private int pageSize;
-    private int naviSize;
-    private int totalPage;
-    private int page;
-    private int beginPage;
-    private int endPage;
-    private boolean showPrev;
-    private boolean showNext;
-
-    // 페이징 목적이 아닌 Read 출력을 위함
-
-    public PageHandler(int totalCnt, int page) {
-        this(page);
-    }
-
-
-    public PageHandler(int page) {
-        this.page = page;
-
-
-
-    }
-
-    public PageHandler(int totalCnt, Integer page, Integer pageSize) {
-    }
-
-
-    @Override
-    public String toString() {
-        return "PageHandler{" +
-                "totalCnt=" + totalCnt +
-                ", pageSize=" + pageSize +
-                ", naveSize=" + naviSize +
-                ", totalPage=" + totalPage +
-                ", page=" + page +
-                ", beginPage=" + beginPage +
-                ", endPage=" + endPage +
-                ", showPrev=" + showPrev +
-                ", showNext=" + showNext +
-                '}';
-    }
-
-
-    public int getTotalCnt() {
-        return totalCnt;
-    }
-
-    public void setTotalCnt(int totalCnt) {
-        this.totalCnt = totalCnt;
-    }
-
-    public int getPageSize() {
-        return pageSize;
-    }
-
-    public void setPageSize(int pageSize) {
-        this.pageSize = pageSize;
-    }
-
-    public int getNaviSize() {
-        return naviSize;
-    }
-
-    public void setNaviSize(int naviSize) {
-        this.naviSize = naviSize;
-    }
-
-    public int getTotalPage() {
-        return totalPage;
-    }
-
-    public void setTotalPage(int totalPage) {
-        this.totalPage = totalPage;
-    }
-
-    public int getPage() {
-        return page;
-    }
-
-    public void setPage(int page) {
-        this.page = page;
-    }
-
-    public int getBeginPage() {
-        return beginPage;
-    }
-
-    public void setBeginPage(int beginPage) {
-        this.beginPage = beginPage;
-    }
-
-    public int getEndPage() {
-        return endPage;
-    }
-
-    public void setEndPage(int endPage) {
-        this.endPage = endPage;
-    }
-
-    public boolean isShowPrev() {
-        return showPrev;
-    }
-
-    public void setShowPrev(boolean showPrev) {
-        this.showPrev = showPrev;
-    }
-
-    public boolean isShowNext() {
-        return showNext;
-    }
-
-    public void setShowNext(boolean showNext) {
-        this.showNext = showNext;
-    }
-}
